+++ src/bano_reader.rs
@@ -1,0 +1,6 @@
+pub fn populate_dict_from_bano_file(file: &str, ispell: &mut ::ispell_wrapper::SpellCheck) {
+    println!("Reading street and city names from {}", file);
+
+    ispell.add_word("L'Haÿ-les-Roses").unwrap();
+}
+
